fix: make CLI call timeout configurable
